+++ src/app/cart-list/cart-list.component.ts
@@ -1,8 +1,7 @@
 import { Component, OnDestroy, OnInit } from '@angular/core';
-import { Observable, Subscription } from 'rxjs';
-import { scan } from 'rxjs/operators';
+import { Observable } from 'rxjs';
 
-import { Product } from '../models';
+import { Product } from '../product-list/models';
 import { CartService } from '../services/cart.service';
 
 @Component({
@@ -10,45 +9,34 @@ import { CartService } from '../services/cart.service';
 	templateUrl: './cart-list.component.html',
 	styleUrls: ['./cart-list.component.scss']
 })
-export class CartListComponent implements OnInit, OnDestroy {
-	cartProducts: Product[];
+export class CartListComponent implements OnInit {
 	cartProductsTotalPrice$: Observable<number>;
-
-	private cartProducts$: Observable<Product[]>;
-	private cartProductsSubscription = Subscription.EMPTY;
+	cartProductsAmount$: Observable<number>;
+	cartEmpty$: Observable<boolean>;
+	cartProducts$: Observable<Product[]>;
 
 	constructor(private cartService: CartService) { }
 
 	ngOnInit(): void {
+		this.cartProductsTotalPrice$ = this.cartService.getCartProductsTotalPriceStream();
+		this.cartProductsAmount$ = this.cartService.getCartProductsAmountStream();
 		this.cartProducts$ = this.cartService.getCartProductListStream();
-		this.cartProductsSubscription = this.cartProducts$
-			.subscribe((products: Product[]) => {
-				this.cartProducts = products
-			})
-		this.cartProductsTotalPrice$ = this.cartProducts$
-			.pipe(
-				scan((accumulator, currentValue: Product[]) => {
-					return currentValue.length ? accumulator + this.getProductsTotalPrice(currentValue) : 0;
-				}, 0)
-			);
+		this.cartEmpty$ = this.cartService.getCartEmptyStream();
 	}
 
 	onClearCartClick(): void {
 		this.cartService.clearProductCart();
 	}
 
-	onRemoveClicked(ev: Product): void {
+	onProductDeleted(ev: Product): void {
 		this.cartService.removeProductFromCart(ev);
 	}
 
-	getProductsTotalPrice(products: Product[]): number {
-		return products.reduce((accumulator, currentValue: Product) => {
-			accumulator += currentValue.price;
-			return accumulator;
-		}, 0);
+	onItemAdded(ev: Product): void {
+		this.cartService.increaseProductAmount(ev, 1);
 	}
 
-	ngOnDestroy(): void {
-		this.cartProductsSubscription.unsubscribe()
+	onItemRemoved(ev: Product): void {
+		this.cartService.decreaseProductAmount(ev, 1);
 	}
 }