+++ src/app/product-list/models/product.model.ts
@@ -5,6 +5,7 @@ export interface Product {
 	category: ProductCategory;
 	isAvailable: boolean;
 	id: string;
+	amountInCart?: number
 }
 
 export type ProductCategory = 'healthcare' | 'electronics' | 'clothing';