+++ changelog.md
@@ -1,6 +1,15 @@
-### Features
+### v. 1.0
 
 * Created **product-list**
 * Created **cart**, added possibility to add items to cart.
 * Implemented **total price count** in cart
 * Added possibility to **remove items** in cart
+
+### v. 1.1
+
+* Created **shared module**
+* Added **angular material** to project
+* Added **total products amount count**
+* Added possibility to edit **product quantity** in cart
+* Added **items highlight** in cart
+* Resolved all comments from previous hometask